Move writer into a new directory
