+++ packages/dd-trace/src/opentracing/tracer.js
@@ -5,7 +5,7 @@ const Tracer = opentracing.Tracer
 const Reference = opentracing.Reference
 const Span = require('./span')
 const SpanContext = require('./span_context')
-const Writer = require('../writer')
+const Writer = require('../agent/writer')
 const Recorder = require('../recorder')
 const Sampler = require('../sampler')
 const PrioritySampler = require('../priority_sampler')
